requirements.txt
aiohttp==3.7.4.post0
alembic==1.5.4
amqp==5.0.5
anyjson==0.3.3
appdirs==1.4.4
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
Babel==2.9.0
billiard==3.6.3.0
black==19.10b0
blinker==1.4
celery==5.0.5
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
coverage==5.5
defusedxml==0.7.1
distlib==0.3.1
dnspython==2.1.0
email-validator==1.1.2
filelock==3.0.12
flake8==3.8.4
flake8-bugbear==20.11.1
flake8-polyfill==1.0.2
Flask==1.1.2
Flask-Admin==1.5.7
Flask-BabelEx==0.9.4
Flask-fillin==0.2
Flask-JWT-Extended==4.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==2.6.0
Flask-OpenID==1.2.5
Flask-Principal==0.4.0
flask-pytest==0.0.5
Flask-SQLAlchemy==2.4.4
Flask-Validator==1.4.2
Flask-WTF==0.14.3
identify==2.2.1
idna==2.10
iniconfig==1.1.1
isbnlib==3.10.3
iso3166==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==5.0.2
lxml==4.6.2
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.10.0
mccabe==0.6.1
multidict==5.1.0
nodeenv==1.5.0
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
pre-commit==2.11.1
prompt-toolkit==3.0.16
psycopg2==2.8.6
py==1.10.0
py-moneyed==0.8.0
pycodestyle==2.6.0
pycountry==20.7.3
pyflakes==2.2.0
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.4.1
redis==3.5.3
regex==2020.11.13
requests==2.25.1
schwifty==2020.9.0
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.23
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
vine==5.0.0
virtualenv==20.4.3
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
yarl==1.6.3
